requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bootstrap4==0.1.0
dj-database-url==0.5.0
Django==4.0.5
django-bootstrap4==22.1
django-clearcache==1.2.1
django-crispy-forms==1.14.0
django-on-heroku==1.1.2
django-registration-redux==2.10
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.1.1
psycopg2-binary==2.9.3
python-decouple==3.6
pytz==2022.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
whitenoise==6.2.0
